fix(studio): address review findings on scheduled-unpublish PR
- Move scheduleAction/cancelScheduleAction and the delete-unpublish
  guard out of routers into page.service.ts/resource.service.ts, per
  the router/service split in server/CLAUDE.md.
- Replace schedulePublishingJob's error.code+message string-matching
  for "already unpublished" with a typed PageAlreadyUnpublishedError.
- Loosen cancelScheduleAction's scheduledAction check to treat a null
  value as Publish, matching the cron's own null-handling convention.
- Add user-facing messages to unpublishPageSchema's validators.
